requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
jinja2==3.1.4
    # via -r requirements.in
markupsafe==2.1.5
    # via jinja2
ops==2.13.0
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   ops
websocket-client==1.8.0
    # via ops
